improvements v2
