+++ indigo-frontend/src/core/services/report-bug.service.ts
@@ -1,12 +1,11 @@
 import { DOCUMENT } from '@angular/common';
 import { inject, Injectable } from '@angular/core';
-import {
-  ReportErrorContext,
-  ReportErrorFormValue,
-} from '@core/components/common/report-error-dialog/report-error-dialog.component';
 import { ApiService } from '@core/services/api.service';
 import { ExperimentDetailService } from '@core/services/experiment/experiment-detail.service';
+import { ReportErrorContext, ReportErrorFormValue } from '@core/types/report-error.i';
 import { Observable } from 'rxjs';
+
+export const INCIDENTS_ENDPOINT = 'incidents';
 
 export interface ReportErrorTechnicalDetails {
   currentUrl: string;
@@ -26,7 +25,7 @@ export class ReportBugService {
   private experimentDetailService = inject(ExperimentDetailService);
 
   submitReport(formValue: ReportErrorFormValue, context?: ReportErrorContext | null): Observable<void> {
-    return this.api.request<void>('post', 'incidents', this.buildPayload(formValue, context));
+    return this.api.request<void>('post', INCIDENTS_ENDPOINT, this.buildPayload(formValue, context));
   }
 
   getTechnicalDetails(context?: ReportErrorContext | null): ReportErrorTechnicalDetails | null {
